[#179193861] Render tabs and filters in mobile accordion
